Finishing auth, working on sharing task

+++ backend/database/migrations/2020_09_15_181144_create_user_task_table.php
@@ -15,8 +15,8 @@ class CreateUserTaskTable extends Migration
     {
         Schema::create('user_task', function (Blueprint $table) {
             $table->id();
-            $table->integer('user_id')->unsigned();
-            $table->integer('task_id')->unsigned();
+            $table->bigInteger('user_id')->unsigned();
+            $table->bigInteger('task_id')->unsigned();
             $table->timestamps();
         });
     }
